Lock and Key Scripts
Here are some basic scripts for opening doors. The Key Holder script must be assigned onto the player model. The Key scripts are assigned on 3d models that are designated as keys to open doors. You will assign it a key type, I have red green and blue but u can add more under the script. You then add the key door script onto a door you wish to be opened, assign it the same key type as the key and it will open. Make sure the key and door 3d objects have a collider that is a trigger.