12/4/2021
Created a rat, rigged and animated it as level enemy. Added sounds into level with scripts and objects to control them such as turnstile clicks and announcements.

Started working on flashing red indicator plane controlled by script. Still need to make a method that turns on red plane and reduces for 2 seconds back to normal. The player script will access this method when health is affected.